requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
Pillow>=10.0
python-decouple>=3.8
django-filter>=23.5
whitenoise>=6.6
gunicorn>=21.2
